server/requirements.txt
aiohttp==3.5.4
PyYAML==5.1